designed rest apis for vehicle status & refactored code.

+++ src/main/java/com/tms/loader/payloads/StatusDto.java
@@ -10,6 +10,6 @@ import lombok.ToString;
 @NoArgsConstructor
 @ToString
 public class StatusDto {
-	private Long statusId;
+	private Integer statusId;
 	private String status;
 }
