+++ src/main/java/com/tms/loader/repositories/vehicle/VehicleTypeRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tms.loader.entities.vehicle.VehicleType;
 
-public interface VehicleTypeRepo extends JpaRepository<VehicleType, Long> {
+public interface VehicleTypeRepo extends JpaRepository<VehicleType, Integer> {
 
 }
